fix migrate

+++ db/migrate/20210917083731_add_profile_to_user.rb
@@ -9,11 +9,23 @@ class AddProfileToUser < ActiveRecord::Migration[6.1]
     add_column :users, :summary, :text
     add_column :users, :highest_education, :string
     add_column :users, :highest_career, :string
-    remove_column :users, :root
-    remove_column :users, :admin
-    remove_column :users, :user
-    remove_column :users, :employer
-    remove_column :users, :company
-    remove_column :users, :company_id
+    if column_exists? :users, :root
+      remove_column :users, :root
+    end
+    if column_exists? :users, :admin
+      remove_column :users, :admin
+    end
+    if column_exists? :users, :user
+      remove_column :users, :user
+    end
+    if column_exists? :users, :employer
+      remove_column :users, :employer
+    end
+    if column_exists? :users, :company
+      remove_column :users, :company
+    end
+    if column_exists? :users, :company_id
+      remove_column :users, :company_id
+    end
   end
 end
